Workaround to correct nested modules dir is now only executed if problem exists

+++ bin/simplesamlphp-post-install.sh
@@ -29,8 +29,10 @@ cp config/* vendor/simplesamlphp/simplesamlphp/config/
 cp metadata/* vendor/simplesamlphp/simplesamlphp/metadata/
 
 # Workaround: move modules back to correct location, this happpens due to incorrect installation order (janus before ssp)
-mv vendor/simplesamlphp/simplesamlphp/modules/modules/* vendor/simplesamlphp/simplesamlphp/modules/
-rm -r vendor/simplesamlphp/simplesamlphp/modules/modules
+if [-d mv vendor/simplesamlphp/simplesamlphp/modules/modules/* vendor/simplesamlphp/simplesamlphp/modules/]; then
+    mv vendor/simplesamlphp/simplesamlphp/modules/modules/* vendor/simplesamlphp/simplesamlphp/modules/
+    rm -r vendor/simplesamlphp/simplesamlphp/modules/modules
+fi
 
 # Enable cron module
 touch vendor/simplesamlphp/simplesamlphp/modules/cron/enable
